refactor: restructured hatchout-core
