+++ contracts/ethereum/packages/hatchout-core/methods/createEggMethod.ts
