+++ contracts/ethereum/packages/hatchout-core/methods/createSaleAuctionMethod.ts
