+++ contracts/ethereum/packages/hatchout-core/methods/createSpecialAuctionMethod.ts
