+++ contracts/ethereum/packages/hatchout-core/methods/getGhostsMethod.ts
