+++ contracts/ethereum/packages/hatchout-core/methods/levelUpMethod.ts
