+++ contracts/ethereum/packages/hatchout-core/package.json
@@ -1,0 +1,57 @@
+{
+  "name": "hatchout-core",
+  "version": "1.0.0",
+  "description": "",
+  "scripts": {},
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "@types/web3": "^1.0.19",
+    "chai": "^4.2.0",
+    "deep-equal-in-any-order": "^1.0.16",
+    "package.json": "^2.0.1",
+    "web3": "^1.0.0-beta.55",
+    "web3-providers": "^1.0.0-beta.55"
+  },
+  "devDependencies": {
+    "@haechi-labs/vvisp-contracts": "^1.0.0",
+    "@haechi-labs/vvisp-utils": "^1.0.0",
+    "@types/jest": "^23.3.13",
+    "@types/node": "^10.12.18",
+    "babel-preset-env": "^1.7.0",
+    "coveralls": "^3.0.1",
+    "deep-eql": "^4.0.0",
+    "eth-gas-reporter": "^0.1.7",
+    "ethereumjs-util": "^5.1.2",
+    "ethjs-abi": "^0.2.1",
+    "ganache-cli": "6.1.0",
+    "husky": "^1.2.1",
+    "lint-staged": "^8.1.0",
+    "lodash": "^4.17.15",
+    "openzeppelin-solidity": "2.1.3",
+    "openzeppelin-test-helpers": "^0.3.2",
+    "prettier": "^1.15.3",
+    "solidity-coverage": "git+https://github.com/rotcivegaf/solidity-coverage.git#5875f5b7bc74d447f3312c9c0e9fc7814b482477",
+    "solium": "^1.1.6",
+    "truffle": "^5.0.0",
+    "ts-node": "^8.3.0",
+    "ts-node-register": "^1.0.0",
+    "typescript": "^3.5.3"
+  },
+  "lint-staged": {
+    "**/*.js": [
+      "prettier --write",
+      "git add"
+    ]
+  },
+  "husky": {
+    "hooks": {
+      "pre-commit": "lint-staged"
+    }
+  },
+  "prettier": {
+    "singleQuote": true,
+    "tabWidth": 2,
+    "printWidth": 80
+  }
+}
