+++ contracts/ethereum/packages/hatchout-core/proxy/methodProxy.ts
