+++ contracts/ethereum/packages/hatchout-core/suppliers/vvispStateSupplier.ts
