+++ contracts/ethereum/packages/hatchout-core/types/index.ts
