+++ contracts/ethereum/packages/hatchout-core/utils/utils.ts
